desktop/requirements.txt
PySide6>=6.8
pyserial>=3.5
pyqtgraph>=0.13.7
pytest>=8.0

